Changed protein to double

+++ Solution.java
@@ -3,7 +3,7 @@ import java.util.ArrayList;
 public class Solution {
 	
 	private ArrayList<Item> included;
-	private int maxProtein;
+	private double maxProtein;
 	private int finalCalCount;
 	
 	public ArrayList<Item> getIncluded() {
@@ -18,18 +18,18 @@ public class Solution {
 	
 	public Solution(){
 		included = new ArrayList<Item>();
-		setMaxProtein(0);
+		setMaxProtein(0.0);
 		setFinalCalCount(0);
 	}
 	
 	
-	public Solution(ArrayList<Item> a, int m, int c){
+	public Solution(ArrayList<Item> a, double m, int c){
 		included = a;
 		setMaxProtein(m);
 		setFinalCalCount(c);
 	}
 	
-	public Solution(int m, int c) {
+	public Solution(double m, int c) {
 		included = new ArrayList<Item>();
 		setMaxProtein(m);
 		setFinalCalCount(c);
@@ -46,12 +46,12 @@ public class Solution {
 	}
 
 
-	public int getMaxProtein() {
+	public double getMaxProtein() {
 		return maxProtein;
 	}
 
 
-	public void setMaxProtein(int maxProtein) {
+	public void setMaxProtein(double maxProtein) {
 		this.maxProtein = maxProtein;
 	}
